# server/build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"EncounterRepl": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"time": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"uid": {
						"type": "string"
					},
					"secure": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"email",
					"location",
					"time",
					"date",
					"uid",
					"secure"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "covidsafeserver",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/encounters": {
			"get": {
				"operationId": "FetchEncounters",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/EncounterRepl"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": ""
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/delete-encounters": {
			"post": {
				"operationId": "DeleteEncounter",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/api/add-encounters": {
			"post": {
				"operationId": "AddEncounter",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/api/send-message": {
			"post": {
				"operationId": "SendMessage",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}